requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-WTF==0.10.2
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
beautifulsoup4==4.3.2
flask-appconfig==0.9.1
flask-mongoengine==0.7.1
itsdangerous==0.24
mongoengine==0.8.7
pymongo==2.7.2
pyparsing==2.0.2
pyswisseph==2.00.00-2
six==1.8.0
svgwrite==1.1.6
wsgiref==0.1.2
